+++ src/error.rs
@@ -61,6 +61,9 @@ pub enum StreamError {
     /// An unknown token.
     InvalidName,
 
+    /// An attribute value must not contain the `<` character.
+    InvalidAttributeValue,
+
     /// An invalid/unexpected character.
     ///
     /// The first byte is an actual one, others - expected.
@@ -98,6 +101,9 @@ impl fmt::Display for StreamError {
             }
             StreamError::InvalidName => {
                 write!(f, "invalid name token")
+            }
+            StreamError::InvalidAttributeValue => {
+                write!(f, "attribute value with '<' character is not allowed")
             }
             StreamError::InvalidChar(ref chars, pos) => {
                 // Vec<u8> -> Vec<String>